fix: update Property entity to use double for Bedrooms and ensure consistent culture info in conversions

+++ Data/Entities/Property.cs
@@ -14,7 +14,7 @@ public partial class Property
 
     public string Currency { get; set; } = null!;
 
-    public int Bedrooms { get; set; }
+    public double Bedrooms { get; set; }
 
     public int Bathrooms { get; set; }
